Using GLMMs to estimate time trends
Nearly all of these response variables are counts, and some of them are overdispersed. This PR uses Poisson GLMMs to estimate county-specific time trends for each response variable. These are provided at the end of `5_county_slopes.R` in an object called `county_slopes`. The values here are the estimated increase in the log expected response for an increase in one year.

For each response, I tried an overdispersed and not overdispersed Poisson GLMM and used the one with the lowest AIC.

Minor changes: I added `dplyr::` before a call to `select` in `1_get_clean_data.R` to avoid an error.